kpet pair 준비 중 2019-03-28 오후 3시

+++ app/src/main/java/com/animal/scale/hodoo/domain/Statistics.java
@@ -17,6 +17,6 @@ public class Statistics implements Serializable {
 
     private String theYear;
 
-    private int average;
+    private float average;
 
 }
